# lab3/docs/swagger.yaml
basePath: /api/v1
definitions:
  serializer.BigORequestJSON:
    properties:
      bigo_request_id:
        type: integer
      calculated_complexity:
        type: string
      calculated_time:
        type: number
      creator_login:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      moderator_login:
        type: string
      status:
        type: string
    type: object
  serializer.CompClassRequestJSON:
    properties:
      array_size:
        type: integer
      big_o_request_id:
        type: integer
      comp_class_request_id:
        type: integer
      complexclass_id:
        type: integer
    type: object
  serializer.ComplexClassJSON:
    properties:
      compclass_id:
        type: integer
      complexity:
        type: string
      degree:
        type: number
      degree_text:
        type: string
      description:
        type: string
      is_delete:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bigorequest.com
    name: API Support
    url: http://localhost:8080
  description: API для управления расчётами времени и сложности Классов сложности
  license:
    name: MIT
  title: Big O Request API
  version: "1.0"
paths:
...
  /compclassrequest/{compclass_id}/{bigo_request_id}:
    delete:
      description: Удаляет связь класса сложности и заявки
      parameters:
      - description: ID класса сложности
        in: path
        name: compclass_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: bigo_request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная заявка
          schema:
            $ref: '#/definitions/serializer.BigORequestJSON'
        "400":
          description: Неверные ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить класс сложности из заявки
      tags:
      - CompClassRequest
    put:
      consumes:
      - application/json
      description: Обновляет параметры класса сложности в конкретной заявке
      parameters:
      - description: ID класса сложности
        in: path
        name: compclass_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: bigo_request_id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/serializer.CompClassRequestJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленные данные
          schema:
            $ref: '#/definitions/serializer.CompClassRequestJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные класса сложности в заявке
      tags:
      - CompClassRequest
  /complexclass:
    get:
      description: Возвращает все классы сложности или фильтрует по степени
      parameters:
      - description: Степень класса сложности для поиска
        in: query
        name: search-degree
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список классов сложности
          schema:
            items:
              $ref: '#/definitions/serializer.ComplexClassJSON'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список классов сложности
      tags:
      - CompClasses
  /complexclass/{id}:
    get:
      description: Возвращает информацию о классе сложности по его идентификатору
      parameters:
      - description: ID класса сложности
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные класса сложности
          schema:
            $ref: '#/definitions/serializer.ComplexClassJSON'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить класс сложности по ID
      tags:
      - CompClasses
  /complexclass/{id}/add-photo:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для класса сложности и возвращает обновленные
        данные
      parameters:
      - description: ID класса сложности
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение класса сложности
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Статус загрузки и данные класса сложности
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос или файл
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Класс сложности не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение устройства
      tags:
      - CompClasses
  /complexclass/{id}/add-to-bigorequest:
    post:
      description: Добавляет класс сложности в заявку-черновик пользователя
      parameters:
      - description: ID класса сложности
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Расчёт с добавленным классом сложности
          schema:
            $ref: '#/definitions/serializer.BigORequestJSON'
        "201":
          description: Создан новый расчёт
          schema:
            $ref: '#/definitions/serializer.BigORequestJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Класс сложности не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Класс сложности уже в расчёте
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить класс сложности в расчёт
      tags:
      - CompClasses
  /complexclass/{id}/delete-compclass:
    delete:
      description: Выполняет логическое удаление класса сложности по ID
      parameters:
      - description: ID класса сложности
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Класс сложности не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить класс сложности
      tags:
      - CompClasses
  /complexclass/{id}/edit-compclass:
    put:
      consumes:
      - application/json
      description: Обновляет информацию о классе сложности по ID
      parameters:
      - description: ID класса сложности
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные класса сложности
        in: body
        name: compclass
        required: true
        schema:
          $ref: '#/definitions/serializer.ComplexClassJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный класс сложности
          schema:
            $ref: '#/definitions/serializer.ComplexClassJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Класс сложности не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные класса сложности
      tags:
      - CompClasses
  /complexclass/create-compclass:
    post:
      consumes:
      - application/json
      description: Создает новый класс сложности и возвращает его данные
      parameters:
      - description: Данные нового класса сложности
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/serializer.ComplexClassJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный класс сложности
          schema:
            $ref: '#/definitions/serializer.ComplexClassJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новый класс сложности
      tags:
      - CompClasses
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"